@@ -1,12 +1,13 @@
 {
   "name": "ias-jsx-react",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
+    "@chiragrupani/fullscreen-react": "^1.2.1",
     "@dropzone-ui/react": "^6.7.1",
     "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
-    "@hms-dbmi/viv": "^0.12.6",
+    "@hms-dbmi/viv": "^0.13.1",
     "@luma.gl/core": "^8.5.10",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
@@ -19,34 +20,47 @@
     "@mui/lab": "^5.0.0-alpha.74",
     "@mui/material": "^5.5.1",
     "@mui/x-data-grid": "^5.7.0",
+    "@ramonak/react-progress-bar": "^4.4.0",
     "@reduxjs/toolkit": "^1.8.0",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
     "bootstrap": "^4.5.2",
+    "buffer": "^6.0.3",
+    "classnames": "^2.3.1",
     "connected-react-router": "^6.9.2",
+    "crypto": "^1.0.1",
     "deck.gl": "^8.7.2",
+    "fs": "^0.0.1-security",
+    "geotiff": "^2.0.5",
     "history": "^5.3.0",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.21",
     "material-ui-search-bar": "^1.0.0",
     "mdi-material-ui": "^7.2.0",
-    "react": "^17.0.2",
+    "path": "^0.12.7",
+    "plotty": "^0.4.9",
+    "react": "^18.1.0",
     "react-bootstrap": "^1.3.0",
-    "react-dom": "^17.0.2",
+    "react-data-grid": "^7.0.0-beta.13",
+    "react-dom": "^18.1.0",
     "react-dropzone": "^11.2.3",
     "react-fontawesome": "^1.7.1",
     "react-icons": "^3.11.0",
+    "react-loader-spinner": "^5.3.4",
+    "react-md": "^5.1.3",
     "react-redux": "^7.2.6",
     "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.0",
     "react-scrollbar": "^0.5.6",
+    "react-step-range-slider": "^1.4.1",
     "redux": "^4.1.2",
     "sass": "^1.49.9",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.1.2",
+    "tiff.js": "^1.0.0",
     "usehooks-ts": "^2.4.2",
+    "utif": "^3.1.0",
     "web-vitals": "^2.1.4",
     "zustand": "^3.4.1"
   },
@@ -73,5 +87,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.5"
   }
 }
