@@ -1,14 +1,17 @@
 from pathlib import Path
 from typing import List
 import os
-import PIL
-from skimage import io
-import numpy as np
 import datetime
 from fastapi import UploadFile
 import aiofiles
+import PIL
+import tifffile
+from skimage import io
+import numpy as np
+
 from motor.motor_asyncio import AsyncIOMotorDatabase
 from mainApi.app.auth.models.user import UserModelDB, PyObjectId, ShowUserModel
+from mainApi.app.images.sub_routers.tile.models import NamePattenModel
 from mainApi.app.images.sub_routers.tile.models import TileModelDB
 from mainApi.app.images.utils.folder import get_user_cache_path, clear_path
 from mainApi.app import main
@@ -19,6 +22,75 @@ async def save_upload_file(upload_file: UploadFile, destination: Path, chunk_siz
         while content := await upload_file.read(chunk_size):  # async read chunk
             await out_file.write(content)  # async write chunk
 
+async def add_image_tiles(path: Path,
+                        files: List[UploadFile],
+                        clear_previous: bool,
+                        current_user: UserModelDB or ShowUserModel,
+                        db: AsyncIOMotorDatabase) -> List[TileModelDB]:
+    """
+    Saves the uploaded tiles to the cache-storage folder/volume under the user_id of the current_user
+
+    Front end should include a validator that checks if the file has already been uploaded and then reject it.
+    No validation is done in the backend
+    """
+    
+    tiles: List[TileModelDB] = []
+    filenames = []
+    for each_file in files:
+        file_path = os.path.join(path, each_file.filename) 
+        async with aiofiles.open(file_path, 'wb') as f:
+            content = await each_file.read()
+            await f.write(content)
+        filenames.append(each_file.filename)
+        width_px, height_px = PIL.Image.open(file_path).size
+        tile = TileModelDB(
+            user_id=PyObjectId(current_user.id),
+            absolute_path=str(path),
+            file_name=each_file.filename,
+            content_type=each_file.content_type,
+            width_px=width_px,
+            height_px=height_px
+        )
+        tiles.append(tile)
+    await db['tile-image-cache'].insert_many([t.dict(exclude={'id'}) for t in tiles])
+    return {"Flag_3d": True, "N_images": len(filenames), "images": filenames}
+    # cache_path = STATIC_PATH
+    # raw_source = io.imread(path, True)
+    # res = raw_source
+    # image_num = res.shape    
+    # current_time = datetime.datetime.now() 
+    # time_str = current_time.strftime("%Y%m%d_%H%M%S")
+    # path_images = []
+    # if image_num[2] > 3:
+    #     for i in range(image_num[0]):
+    #         path_image = os.path.join(cache_path, 'slice_{num:03d}'.format(num=i)+ '_' + time_str + '.png')
+    #         io.imsave(path_image, normalize_2Dim_uint8(res[i]))
+    #         path_image = path_image.split('/')[-1]
+    #         path_images.append(path_image)
+    #     D_flag = True
+    #     return D_flag, image_num[0], path_images
+    # else:
+        # tiles: List[TileModelDB] = []
+        # file = files[0]
+        # path_image = os.path.join(cache_path, 'slice_000_'+ time_str +'.png')
+        # io.imsave(path_image, normalize_2Dim_uint8(res))
+        # path_image = path_image.split('/')[-1]
+        # path_images.append(path_image)
+        # width_px, height_px = PIL.Image.open(file.file).size
+        # tile = TileModelDB(
+        #     user_id=PyObjectId(current_user.id),
+        #     absolute_path=str(path),
+        #     file_name=file.filename,
+        #     content_type=file.content_type,
+        #     width_px=width_px,
+        #     height_px=height_px
+        # )
+        # tiles.append(tile)
+        # await db['tile-image-cache'].insert_many([t.dict(exclude={'id'}) for t in tiles])
+        # path = tiles[0].absolute_path
+        # D_flag = False
+        # image_num = 1
+        # return D_flag, image_num, path_images
 
 def convol2D_processing(file_name):
     abs_path = STATIC_PATH
@@ -37,59 +109,3 @@ def normalize_2Dim_uint8(im):
     max = np.max(im)
     im = (im-min)/(max-min)*255
     return im.astype(np.uint8)
-
-async def add_image_tiles(path: Path,
-                            files: List[UploadFile],
-                          clear_previous: bool,
-                          current_user: UserModelDB or ShowUserModel,
-                          db: AsyncIOMotorDatabase) -> List[TileModelDB]:
-    """
-    Saves the uploaded tiles to the cache-storage folder/volume under the user_id of the current_user
-
-    Front end should include a validator that checks if the file has already been uploaded and then reject it.
-    No validation is done in the backend
-    """
-    cache_path = STATIC_PATH
-    if not os.path.exists(cache_path):
-        os.makedirs(cache_path)
-    raw_source = io.imread(path)
-    res = raw_source
-    image_num = res.shape
-    
-    current_time = datetime.datetime.now() 
-    time_str = current_time.strftime("%Y%m%d_%H%M%S")
-    path_images = []
-
-    if image_num[2] > 3:
-        for i in range(image_num[0]):
-            path_image = os.path.join(cache_path, 'slice_{num:03d}'.format(num=i)+ '_' + time_str + '.png')
-            io.imsave(path_image, normalize_2Dim_uint8(res[i]))
-            path_image = path_image.split('/')[-1]
-            path_images.append(path_image)
-        D_flag = True
-        return D_flag, image_num[0], path_images
-
-    else:
-        tiles: List[TileModelDB] = []
-        file = files[0]
-        path_image = os.path.join(cache_path, 'slice_000_'+ time_str +'.png')
-        io.imsave(path_image, normalize_2Dim_uint8(res))
-        path_image = path_image.split('/')[-1]
-        path_images.append(path_image)
-        width_px, height_px = PIL.Image.open(file.file).size
-
-        tile = TileModelDB(
-            user_id=PyObjectId(current_user.id),
-            absolute_path=str(path),
-            file_name=file.filename,
-            content_type=file.content_type,
-            width_px=width_px,
-            height_px=height_px
-        )
-        tiles.append(tile)
-
-        await db['tile-image-cache'].insert_many([t.dict(exclude={'id'}) for t in tiles])
-        path = tiles[0].absolute_path
-        D_flag = False
-        image_num = 1
-        return D_flag, image_num, path_images
