@@ -6,13 +6,12 @@ import { api } from "./base";
 // COLOR_CHANNEL: `${API_URL}color-channel`,
 // CHANGE_PARAMETER: `${API_URL}change-parameter`,
 // GRAY: `${API_URL}gray`,
-
 export const uploadExperimentData = params => {
-    
-  console.log(params)
-  return api.post("/experiment/upload_experiment_data", params);
+
+    console.log(params)
+    return api.post("/experiment/upload_experiment_data", params);
 };
 
 export const getExperimentData = params => {
-  return api.post("/experiment/get_experiment_data");
+    return api.post("/experiment/get_experiment_data");
 }
