@@ -8,42 +8,37 @@ import { api } from "./base";
 // GRAY: `${API_URL}gray`,
 
 export const setImage = params => {
-  return api.post("set-image", params);
+    return api.post("set-image", params);
 };
 
 export const fileUpload = params => {
-  console.log(params)
-  return api.post("/image/tile/upload_image_tiles", params);
-};
-
-export const setMetadata = params => {
-  console.log(params)
-  return api.post("/image/tile/upload_image_tiles", params);
+    console.log(params)
+    return api.post("/image/tile/upload_image_tiles", params);
 };
 
 export const deconvol2D = params => {
-  console.log(params);
-  return api.post("/image/tile/deconvol2D", params);
+    console.log(params);
+    return api.post("/image/tile/deconvol2D", params);
 };
 
 export const deconvol3D = params => {
-  console.log(params);
-  return api.post("/image/tile/deconvol3D", params);
+    console.log(params);
+    return api.post("/image/tile/deconvol3D", params);
 };
 
 export const sendImageFile_superresolution = params => {
-  console.log(params);
-  return api.post("/image/tile/SuperRes", params);
+    console.log(params);
+    return api.post("/image/tile/SuperRes", params);
 };
 
 export const changeImage = params => {
-  return api.post("change-image", params);
+    return api.post("change-image", params);
 };
 
 export const changeParameter = params => {
-  return api.post("change-parameter", params);
+    return api.post("change-parameter", params);
 };
 
 export const adjustImage = params => {
-  return api.post("adjust-image", params);
+    return api.post("adjust-image", params);
 };
