@@ -12,9 +12,7 @@ import WellPlates from './WellPlates';
 import Dishes from './Dishes';
 import Wafers from './Wafers';
 import { getVesselById } from '../../utils/vessel-types';
-
-import { useElementSize } from 'usehooks-ts';
-
+// import { useElementSize } from 'usehooks-ts';
 
 function TabPanel(props) {
     const { children, value, index, ...other } = props;
@@ -44,20 +42,15 @@ TabPanel.propTypes = {
 
 export const ExpansionDialog = (props) => {
 
-    const [tab, setTab] = useState(0);
+
+    const maxDialogWidth = 600
+    // const [tab, setTab] = useState(0);
     const [open, setOpen] = useState(true);
     const [currentVessel, setCurrentVessel] = useState(props.currentVessel);
-
-    const [ref, { width, height }] = useElementSize();
-
-    useEffect(() => {
-        setOpen(props.open);
-        setCurrentVessel(props.currentVessel);
-    }, [props]);
-
-    const handleTabChange = (event, newValue) => {
-        setTab(newValue);
-    };
+    // const [ref, { width, height }] = useElementSize();
+    // const handleTabChange = (event, newValue) => {
+    //     setTab(newValue);
+    // };
 
     const handleClose = () => {
         props.closeDialog();
@@ -70,26 +63,35 @@ export const ExpansionDialog = (props) => {
         }
     };
 
+    useEffect(() => {
+        setOpen(props.open);
+        setCurrentVessel(props.currentVessel);
+    }, [props]);
+
     const renderVesselItem = (vessel) => {
         if (vessel) {
             switch (vessel.type) {
                 case 'Slide':
-                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: 700 }}><Slides width={700} count={vessel.count} key={vessel.id} /></div>;
-                case 'WellPlate':
-                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: 700 }}><WellPlates width={700} rows={vessel.rows} cols={vessel.cols} showName={true} key={vessel.id} showNumber={true} /></div>;
+                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: maxDialogWidth }}><Slides width={maxDialogWidth} count={vessel.count} key={vessel.id} /></div>;
                 case 'Dish':
-                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: 700 }}><Dishes width={700} size={vessel.size} key={vessel.id} /></div>;
+                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: maxDialogWidth }}><Dishes width={maxDialogWidth} size={vessel.size} key={vessel.id} /></div>;
+                case 'WellPlate':
+                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: maxDialogWidth }}><WellPlates width={maxDialogWidth} rows={vessel.rows} cols={vessel.cols} showName={true} key={vessel.id} showNumber={true} /></div>;
                 case 'Wafer':
-                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: 700 }}><Wafers width={700} size={vessel.size} key={vessel.id} /></div>;
+                    return <div role="button" onClick={() => { changeCurrentVessel(vessel.id) }} style={{ width: maxDialogWidth }}><Wafers width={maxDialogWidth} size={vessel.size} key={vessel.id} /></div>;
+                default:
+                    return;
             }
         }
     }
 
     return (
-        <Dialog open={open} onClose={handleClose} maxWidth={'700'}>
-            <DialogTitle>Vessel Expansion</DialogTitle>
-
-            <div style={{ width: 700 }}>
+        <Dialog open={open} onClose={handleClose} maxWidth={(maxDialogWidth + 5).toString()}>
+            <div className="d-flex border-bottom">
+                <DialogTitle>Vessel Expansion</DialogTitle>
+                <button className="dialog-close-btn" color="primary" size="small" onClick={handleClose}>&times;</button>
+            </div>
+            <div style={{ width: maxDialogWidth + 5 }} className='border'>
                 {renderVesselItem(getVesselById(currentVessel))}
             </div>
             <DialogActions>
