@@ -2,16 +2,16 @@ import React from 'react';
 import { useDropzone } from '../hooks';
 
 export default function DropzoneWrapper({ children }) {
-  const { getRootProps, getInputProps } = useDropzone();
+    const { getRootProps, getInputProps } = useDropzone();
 
-  return (
-    <div
-      /* eslint-disable-next-line react/jsx-props-no-spreading */
-      {...getRootProps({ onClick: event => event.stopPropagation() })}
-    >
-      {/* eslint-disable-next-line react/jsx-props-no-spreading */}
-      <input {...getInputProps()} />
-      {children}
-    </div>
-  );
+    return (
+        <div
+            /* eslint-disable-next-line react/jsx-props-no-spreading */
+            {...getRootProps({ onClick: event => event.stopPropagation() })}
+        >
+            {/* eslint-disable-next-line react/jsx-props-no-spreading */}
+            <input {...getInputProps()} />
+            {children}
+        </div>
+    );
 }
