@@ -1,8 +1,8 @@
-const experiment = (state = 0, action)  =>{
-  switch (action.type) {
-    default:
-      return state
-  }
+const experiment = (state = 0, action) => {
+    switch (action.type) {
+        default:
+            return state
+    }
 };
 
 export default experiment;
